Aligne verticalement le lien "indicateur précédent" avec le titre de l'indicateur et extrait le composant dans son propre fichier
