+++ app.territoiresentransitions.react/src/app/pages/collectivite/Indicateurs/detail/HeaderIndicateur.tsx
@@ -1,7 +1,6 @@
 import classNames from 'classnames';
-import {Link} from 'react-router-dom';
 import HeaderTitle from 'ui/HeaderTitle';
-import {usePrevAndNextIndicateurLinks} from './usePrevAndNextIndicateurLinks';
+import {IndicateurTopNav} from './IndicateurTopNav';
 
 /** Affiche l'en-tête d'une page détail d'un indicateur */
 type HeaderProps = {
@@ -21,7 +20,7 @@ export const HeaderIndicateur = ({
       <HeaderTitle
         titre={title}
         customClass={{
-          container: classNames('!bg-bf925 sticky top-0 z-40', {
+          container: classNames('!bg-bf925 sticky top-0 z-40 px-0 fr-px-7v', {
             'pb-0': !readonly,
           }),
           text: classNames('!text-[#3a3a3a] text-[1rem]', {'pb-4': !readonly}),
@@ -32,35 +31,4 @@ export const HeaderIndicateur = ({
       <IndicateurTopNav />
     </>
   );
-};
-
-/**
- * Affiche les boutons "Indicateur précédent" et "Indicateur suivant" en haut de page
- */
-export const IndicateurTopNav = () => {
-  const {prevIndicateurLink, nextIndicateurLink} =
-    usePrevAndNextIndicateurLinks();
-
-  return prevIndicateurLink || nextIndicateurLink ? (
-    <div className="!bg-bf925 min-h-[1.5rem] flex justify-between fr-text--sm !m-0 fr-pb-3w fr-px-5w overflow-hidden">
-      {prevIndicateurLink ? (
-        <Link
-          to={prevIndicateurLink}
-          className="fr-fi-arrow-left-line fr-btn--icon-left active-transparent"
-        >
-          Indicateur précédent
-        </Link>
-      ) : (
-        <div />
-      )}
-      {nextIndicateurLink && (
-        <Link
-          to={nextIndicateurLink}
-          className="justify-self-end fr-fi-arrow-right-line fr-btn--icon-right active-transparent"
-        >
-          Indicateur suivant
-        </Link>
-      )}
-    </div>
-  ) : null;
 };
